Large name change in the control flow graph module.
Found out that the code does not function correctly. Breaks and continues are not processed correctly, leading to invalid control flow graphs. This in turn leads to all other graphs to be invalid.

I may have to abandon this code base and start from scratch. It is impossible to determine what code is reliable and what is malfunctioning, as the code is hard to read and the test suite is also invalid.