Array access problem is not yet fully fixed, as it runs into problems when the array is a global variable. It is partially fixed but need a better solution.
Changed the method names to solve a conflict when there are two equally named methods in a file. This causes problems when rendering a CFG, as there is a box conflict.

Added support for empty bodies on Cases and Catches. An EmptyCF() is returned and a conditional will avoid problematic constructs.

Added support for try-finally blocks, where there are no catch clauses. Previously it would not process correctly and in some cases there would be missing edges to the exit node.

May have to inline a final block for every exit node, as otherwise there will be invalid control flows. This caused problems for some function processing though. Will have to investigate.

Have to investigate an issue when there is a utility function to throw exceptions. This is not registered as a throw and thus leads to an invalid control flow. No idea yet as to how to fix this problem.